Internal Keycloak setup for Ovirt Engine SSO (#10)
* Create initial admin user

This patch adds ability to create keycloak initial admin user as described by [1].
Username is 'admin' and password same as one set for ovirt engine administrator panel during engine-setup.

[1] https://www.keycloak.org/docs/latest/server_admin/index.html#creating-the-account-remotely

Signed-off-by: Artur Socha <asocha@redhat.com>

* Setup for internal sso(openidc)

This patch prepares Keycloak realm and client for ovirt engine internal
sso needs.
This includes:
- creating dedicated Keycloak realm, client and necessary scopes.
- creating Keycloak internal admin user that mimics already existing admin@internal
- preparing ovirt engine extension configuration files
- preparing httpd openidc configuration

Very important assumption is that keycloak based internal sso is enabled
only for *new* and *RPM* based deployment of ovirt-engine. That means that
authentication for development mode stays as it is as well as for
upgraded installations.

Future steps (outside of the scope of this patch) includes:
- replacing jdbc based internal sso for non-dev environments
- preparing standalone tool for jdbc users migration for upgrading
  existing installation.

Signed-off-by: Artur Socha <asocha@redhat.com>

* Keycloak H2 db replaced with existing posgresql

This patch includes:
- plugins to replace keycloak datasource configuration macros in
  Wildfly/EAP configuration
- Plugin to create dedicated keycloak database. Local or remote db are supported similarly to engine's db

Things to be done in following patches:
- backup / rollback

Signed-off-by: Artur Socha <asocha@redhat.com>

* Keycloak database backup & clearing

This patch add a support to backup keycloak database and clear it for
engine-cleanup.

Signed-off-by: Artur Socha <asocha@redhat.com>

* Add package to versionlock

Signed-off-by: Artur Socha <asocha@redhat.com>

* No keycloak setup for existing environments

Do not setup keycloak and its database for existing (legacy)
environments. Keycloak setup can be performed manually if needed.
One of the reason is the need to migrate existing JDBC or LDAP user
base which is currently not automated (yet).

Signed-off-by: Artur Socha <asocha@redhat.com>

* Print username for keycloak sso and admin

This is to provide basic intructions on close-up stage how to login
to ovirt engine administrator panel using keycloak sso and how to access
keycloak admin console.

Unfortunately, the 'legacy' aaa instructions are still printed. They
will be valid when user manually reverts keycloak sso to jdbc based user
authentication.
We need to think about a better way to compile that
message.

example:
(...)
[ INFO  ] Restarting ovirt-vmconsole proxy service
          Please use the user 'admin' and password specified in order to login to Keycloak admin console
          Please use the user 'admin@ovirt' and password specified in order to login using Keycloak SSO
          Please use the user 'admin@internal' and password specified in order to login
          Web access is enabled at:
              http://dev.dom:80/ovirt-engine
              https://dev.dom:443/ovirt-engine
(...)

Signed-off-by: Artur Socha <asocha@redhat.com>

* Setup keycloak credentials for ovn provider config

In order to reuse existing ovir ovn provider setup code the information
about Keycloak and its user credentials must be available for
ovirtproviderovn plugin. To allow that this data is registered in
environment.

Signed-off-by: Artur Socha <asocha@redhat.com>

* post-review update

This is a combined list of minor and straightforwards updates suggested
during the code review.

The following patches might contain more non-trivial updates.

Signed-off-by: Artur Socha <asocha@redhat.com>

* Use stdin to provide password

This patch changes how keycloak scripts are used with credentials.
Now, password is passed using stdin.
This removes plain text passwords from the logs and from 'ps' output

Signed-off-by: Artur Socha <asocha@redhat.com>

* Generate OIDCCryptoPassphrase with urandom

Signed-off-by: Artur Socha <asocha@redhat.com>

* Use JBOSS_HOME env for keycloak scripts

Signed-off-by: Artur Socha <asocha@redhat.com>

* Better setup customization

This patch includes:
- disable preferred keycloak authentication and revert to deprecated AAA
- allow to specify password for keycloak admin and admin@ovirt (for ovirt
  administration panel)
- improves setup summary message by displaying keycloak related
  information
- reverts the code to pass password to add keycloak user script
  invocation (the tool has a bug that got manifested itself)

Known bugs:
- single-sign-on to Grafana does not yet work
- for every subsequent engine-setup run user is prompted whether to
  install Keycloak authentication
- engine-setup --reconfigure-optional-components is not able to revert
  from previously selected Keycloak authentication to deprecated AAA
  modules
- When Keycloak authentication is disabled on clean setup and later on
  re-enabled (the following engine-setup run), ovirt-provider-ovn is
  misconfigured. To fix it login to administration panel and modify ovn
  provider by replacing originally selected user and (typically
  admin@local) with admin@ovirt@internalsso and parssword choosen for
  keycloak authentication (if different from AAA's admin@internal)

Signed-off-by: Artur Socha <asocha@redhat.com>

* Support for Grafana SSO

This patch provides information necessary to setup single sign on for
ovirt-dwh grafana (monitoring).

Things to consider:
- enable new user sign-up becuse now the roles can be managed from
  Keycloak, no explicitely assigned role will prevent user even from
  'viewer'.

Signed-off-by: Artur Socha <asocha@redhat.com>

* Keyclaok backup dir in setup package

Signed-off-by: Artur Socha <asocha@redhat.com>

* Readme: gerrit references removed

To trigger build

Signed-off-by: Artur Socha <asocha@redhat.com>

* bump rpm release version

Signed-off-by: Artur Socha <asocha@redhat.com>

* Disable FIPS for adding keycloak user

Unfortunately Keycloak 15.0.2 is not fully supporting FIPS. For more detail
check the following issue comments:
https://github.com/keycloak/keycloak/issues/9916

Signed-off-by: Artur Socha <asocha@redhat.com>

* Allow for seamless engine backup/restore flow

Please not that keycloak db backup/restore is not supported yet.
Will be added in future.

Signed-off-by: Artur Socha <asocha@redhat.com>

* Add email for admin@ovirt user

Signed-off-by: Artur Socha <asocha@redhat.com>