Fixed internal server error when enabling keycloak on existing installation
When you rerun engine-setup to activate Keycloak, it can't find the OVESETUP_CONFIG/adminPassword value in the environmental variables. This becomes a problem when you're setting up Keycloak alone because it needs that password to finish the setup. Since we can't fetch the password in plain text from anywhere, we ask the user to enter the engine password so the setup can complete successfully.

Note: This is not an issue when setting up a fresh installation with engine-setup (and immediately activating keycloak), because the password is stored in the environmental variables.
Signed-off-by: Brooklyn Dewolf <contact@brooklyn.gent>